+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## v1.2.4
+#### 🚀 Features
+
+- Use llamafile v0.6.2 (PR: #111)
+
+#### 🐛 Fixes
+
+- use Debug.LogError instead of Exception for more verbosity (PR: #113)
+
+#### 📦 General
+
+- Removed duplicate EditorGUI.EndChangeCheck() (PR: #110)
+
+
 ## v1.2.3
 #### 🐛 Fixes